Separated the logic of launch and operation of the bot.
