+++ manage.py
@@ -1,21 +1,12 @@
 #!/usr/bin/env python
 """Django's command-line utility for administrative tasks."""
 import os
-
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'confectioner_bot.settings')
-import django
-from django.conf import settings
-
-if not settings.configured:
-    django.setup()
-
 import sys
-from bot.management.commands.bot import cake_bot
 
 
 def main():
     """Run administrative tasks."""
-#    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'confectioner_bot.settings')
+    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'confectioner_bot.settings')
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc:
@@ -24,10 +15,7 @@ def main():
             "available on your PYTHONPATH environment variable? Did you "
             "forget to activate a virtual environment?"
         ) from exc
-    cake_bot()
-
     execute_from_command_line(sys.argv)
-
 
 
 if __name__ == '__main__':
